requirements.txt
Django==5.2.4
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
gunicorn==23.0.0
requests==2.32.5
icecream==2.1.8
whitenoise==6.11.0
